Removed the unused web api tests. Updated build scripts to minify the web api in production. Minor coding style changes.

+++ web_api/build/make/build.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 pushd ../../
-rm -rf library
+if [ $# -eq 0 ] || [ "$1" != "Debug" ]; then export PROD_ENV=1; fi
+if [ -d library ]; then rm -rf library; fi
 node ./node_modules/webpack/bin/webpack.js
 popd
